make displayError accessible from all pages
