+++ templates/layouts/base.html
@@ -12,13 +12,14 @@
     {% endblock %}
   </head>
   <body>
-    
+
     {% block content %}
     
     {% endblock %}
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+    <script src="/static/js/base.js"></script>
 
     {% block scriptContent %}
