+++ templates/layouts/org-layout.html
@@ -27,7 +27,8 @@
       </div>
     </div>
   </nav>
-
+  <div class="alert-container"></div>
+  
   {% block contentInner %}
   
   {% endblock %}
