+++ templates/layouts/vol-layout.html
@@ -29,7 +29,8 @@
       </div>
     </div>
   </nav>
-
+  <div class="alert-container"></div>
+  
   {% block contentInner %}
   
   {% endblock %}
