+++ templates/organization/login.html
@@ -8,6 +8,7 @@
 
 {% block content %}
 
+<div class="alert-container"></div>
 <div class="container-outer">
   <div class="container">
     <div class="row">
@@ -15,7 +16,6 @@
         <a class="pull-right" href="/">Return to homepage</a>
       </div>
     </div>
-    <div class="alert-container"></div>
     <div class="row">
       <h2>Organization Sign-Up</h2>
     </div>
